No Callbacks After Complete
S3_initiate_multipart and S3_complete_multipart_upload reported data
parsed from the response XML after firing the callback indicating the
operation was complete. This made tracking the lifetime of the operation
more difficult than necessary and also broke with the pattern
established by other operations in the library.

Reordered so that complete is the last callback fired in the lifecycle
of these operations.