Request Payer
When accessing "Requester Pays" buckets a special header must be sent to
verify that the requester acknowledges that they will be charged for the
request:

  x-amz-request-payer: requester

See: https://docs.aws.amazon.com/AmazonS3/latest/userguide/ObjectsinRequesterPaysBuckets.html

Previously this was not supported. Added the requestPayer member to
S3BucketContext and corresponding integration to enable this workflow.