Skip Properties Callback on Failure
When a request ends an attempt is made to dispatch the "properties
callback" (which receives information provided by the HTTP headers).
Previously this step was not skipped in a situation where:

- The HTTP headers had been received, and
- Receiving the request body had failed

This had the effect of overwriting the status of the overall request (a
failure) with the status returned by the properties callback (possibly a
success). This meant that requests which failed could be reported as
having succeeded in the above-described scenario.